+++ src/main/java/com/bhaskar/questionservice/repository/QuestionRepo.java
@@ -1,4 +1,4 @@
-package com.bhaskar.questionservice.dao;
+package com.bhaskar.questionservice.repository;
 
 
 import com.bhaskar.questionservice.model.Question;
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface QuestionDao extends JpaRepository<Question, Integer> {
+public interface QuestionRepo extends JpaRepository<Question, Integer> {
 
 
     List<Question> findByCategory(String category);
